requirements.txt
rosbags
bagpy
pandas
cv-bridge
opencv-python
tqdm